server/requirements.txt
asgiref==3.11.1
Django==6.0.3
djangorestframework==3.16.1
gunicorn==25.1.0
packaging==26.0
psycopg==3.3.3
psycopg-binary==3.3.3
sqlparse==0.5.5
typing_extensions==4.15.0
